+++ src/main/scala/com/github/nikdon/telepooz/model/methods/SetWebhook.scala
@@ -1,0 +1,3 @@
+package com.github.nikdon.telepooz.model.methods
+
+case class SetWebhook(url: String /*, certificate: InputFile */)